feat: adding new endpoints for search and delete by currentUser and userId and standardization of endpoints syntax

+++ backend/spring-boot/src/main/java/com/project/cinescope/application/request/impl/RequestServiceImpl.java
@@ -1,5 +1,6 @@
 package com.project.cinescope.application.request.impl;
 
+import com.project.cinescope.application.auth.service.AuthenticatedUserService;
 import com.project.cinescope.application.movie.Movie;
 import com.project.cinescope.application.movie.MovieRepository;
 import com.project.cinescope.application.request.Request;
@@ -9,6 +10,7 @@ import com.project.cinescope.application.request.request.RequestMovieRequestDto;
 import com.project.cinescope.application.request.response.RequestMovieResponseDto;
 import com.project.cinescope.application.user.User;
 import com.project.cinescope.application.user.UserRepository;
+import com.project.cinescope.core.exception.exceptions.ForbiddenOperationException;
 import com.project.cinescope.core.exception.exceptions.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
@@ -18,13 +20,15 @@ import java.util.List;
 public class RequestServiceImpl implements RequestService {
 
     private final RequestRepository requestRepository;
-    private final UserRepository userRepository;
     private final MovieRepository movieRepository;
+    private final UserRepository userRepository;
+    private final AuthenticatedUserService authenticatedUserService;
 
-    public RequestServiceImpl(RequestRepository requestRepository, UserRepository userRepository, MovieRepository movieRepository) {
+    public RequestServiceImpl(RequestRepository requestRepository, MovieRepository movieRepository, UserRepository userRepository, AuthenticatedUserService authenticatedUserService) {
         this.requestRepository = requestRepository;
-        this.userRepository = userRepository;
         this.movieRepository = movieRepository;
+        this.userRepository = userRepository;
+        this.authenticatedUserService = authenticatedUserService;
     }
 
     public List<RequestMovieResponseDto> getAll() {
@@ -40,9 +44,25 @@ public class RequestServiceImpl implements RequestService {
                 .orElseThrow(() -> new ResourceNotFoundException("Request not found with id: " + id));
     }
 
+    public List<RequestMovieResponseDto> findRequestsByCurrentUser() {
+        User user = authenticatedUserService.getCurrentUser();
+        List<Request> requestList = requestRepository.findByUser(user);
+        return requestList.stream()
+                .map(RequestMovieResponseDto::toRequestDto)
+                .toList();
+    }
+
+    public List<RequestMovieResponseDto> findRequestsByUserId(Long userId) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + userId));
+        List<Request> requestList = requestRepository.findByUser(user);
+        return requestList.stream()
+                .map(RequestMovieResponseDto::toRequestDto)
+                .toList();
+    }
+
     public RequestMovieResponseDto post(RequestMovieRequestDto requestDto) {
-        User user = userRepository.findById(requestDto.userId())
-                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + requestDto.userId()));
+        User user = authenticatedUserService.getCurrentUser();
 
         Movie movie = movieRepository.findById(requestDto.movieId())
                 .orElseThrow(() -> new ResourceNotFoundException("Movie not found with id: " + requestDto.movieId()));
@@ -58,6 +78,18 @@ public class RequestServiceImpl implements RequestService {
         Request request = requestRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Request not found with id: " + id));
 
+        requestRepository.delete(request);
+    }
+
+    public void deleteRequestByCurrentUser(Long id) {
+        User user = authenticatedUserService.getCurrentUser();
+
+        Request request = requestRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Request not found with id: " + id));
+
+        if (user != request.getUser()) {
+            throw new ForbiddenOperationException("You don't have permission to delete another user's requests.");
+        }
         requestRepository.delete(request);
     }
 }
